+++ changelog.md
@@ -17,3 +17,4 @@
 - Added Vanish Config flag to disable fake join/leave messages
 - Allow Regex matching for Message filtering
 - Allow message filtering to be used on Usernames as well
+- Access Control/Optional Verification can now give multiple `verifiedRole`s